backend/requirements.txt
# Core dependencies for our FastAPI backend
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1